Extract isArguments to own file
