+++ lib/samsam.js
@@ -2,6 +2,7 @@
 
 var o = Object.prototype;
 var getClass = require("./get-class");
+var isArguments = require("./is-arguments");
 var isDate = require("./is-date");
 var isElement = require("./is-element");
 
@@ -13,28 +14,6 @@ function isNaN(value) {
     return typeof value === "number" && value !== val;
 }
 
-/**
- * @name samsam.isArguments
- * @param Object object
- *
- * Returns ``true`` if ``object`` is an ``arguments`` object,
- * ``false`` otherwise.
- */
-function isArguments(object) {
-    if (getClass(object) === "Arguments") { return true; }
-    if (typeof object !== "object" || typeof object.length !== "number" ||
-            getClass(object) === "Array") {
-        return false;
-    }
-    if (typeof object.callee === "function") { return true; }
-    try {
-        object[object.length] = 6;
-        delete object[object.length];
-    } catch (e) {
-        return true;
-    }
-    return false;
-}
 
 /**
  * @name samsam.isNegZero
